command-line-app/requirements.txt
numpy
scipy
torch
torchvision
matplotlib
Pillow
tqdm